+++ Classes/EventListener/AddBreadcrumbList.php
@@ -25,7 +25,6 @@ use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
 final class AddBreadcrumbList
 {
     private const DEFAULT_DOKTYPES_TO_EXCLUDE = [
-        255, // Remove when compatibility with TYPO3 v12 is dropped
         PageRepository::DOKTYPE_SPACER,
         PageRepository::DOKTYPE_SYSFOLDER,
     ];