+++ ext_emconf.php
@@ -10,8 +10,7 @@ $EM_CONF[$_EXTKEY] = [
     'version' => '4.0.0-dev',
     'constraints' => [
         'depends' => [
-            'php' => '8.1.0-0.0.0',
-            'typo3' => '12.4.0-13.4.99',
+            'typo3' => '13.4.0-13.4.99',
         ],
         'suggests' => [
             'adminpanel' => '',